+++ code/go/0chain.net/sharder/protocol_block_main.go
@@ -1,0 +1,10 @@
+//go:build !integration_tests
+// +build !integration_tests
+
+package sharder
+
+import "0chain.net/chaincore/block"
+
+func notifyConductor(block *block.Block) error {
+	return nil
+}
