+++ code/go/0chain.net/sharder/sharder/sharder_integration_tests.go
@@ -4,14 +4,9 @@
 package main
 
 import (
-	"go.uber.org/zap"
-
 	"0chain.net/chaincore/block"
 	"0chain.net/chaincore/chain"
-	"0chain.net/chaincore/node"
 	crpc "0chain.net/conductor/conductrpc" // integration tests
-	"0chain.net/conductor/conductrpc/stats"
-	"github.com/0chain/common/core/logging"
 )
 
 // start lock, where the sharder is ready to connect to blockchain (BC)
@@ -37,20 +32,3 @@ func readMagicBlock(magicBlockConfig string) (*block.MagicBlock, error) {
 	return chain.ReadMagicBlockFile(magicBlockConfig)
 }
 
-func notifyConductor(block *block.Block) error {
-	logging.Logger.Debug("[conductor] notifyConductor",
-		zap.String("sharder", node.Self.ID),
-		zap.String("miner", block.MinerID),
-		zap.Int64("round", block.Round),
-		zap.String("hash", block.Hash),
-	)
-	if crpc.Client().State().NotifyOnBlockGeneration {
-		return crpc.Client().NotifyOnSharderBlock(&stats.BlockFromSharder{
-			Round: block.Round,
-			Hash: block.Hash,
-			GeneratorId: block.MinerID,
-			SenderId: node.Self.ID,
-		})
-	}
-	return nil
-}
