+++ code/go/0chain.net/sharder/sharder/sharder_main.go
@@ -15,7 +15,3 @@ func shutdownIntegrationTests()     {}
 func readMagicBlock(magicBlockConfig string) (*block.MagicBlock, error) {
 	return chain.ReadMagicBlockFile(magicBlockConfig)
 }
-
-func notifyConductor(block *block.Block) error {
-	return nil
-}
