+++ docker.local/config/conductor.sharders.yaml
@@ -60,25 +60,23 @@ tests:
   
   - name: "Down sharder coming up in different times"
     flow:
-      # - magic_block_config: config/b0magicBlock_3_miners_3_sharders.json
-      # - magic_block_config: config/b0magicBlock_4_miners_2_sharders.json
       - set_monitor: "sharder-1"
       - cleanup_bc: {}
-      - start: ["sharder-1", "sharder-2", "sharder-3"]
+      - start: ["sharder-1", "sharder-2"]
       - start: ["miner-1", "miner-2", "miner-3"]
       - wait_round:
           shift: 100
-      - stop: ["miner-1", "miner-2", "miner-3", "sharder-2", "sharder-3"]
-      - sleep: "2m"
-      - start: ["sharder-2"] # WIll get the old LFB
+      - stop: ["sharder-2"]
+      - wait_round:
+          shift: 100
+      - stop: ["miner-1", "miner-2", "miner-3", "sharder-1"]
       - sleep: "1m"
-      - start: ["miner-1", "miner-2", "miner-3"]
+      - start: ["miner-1", "miner-2", "miner-3", "sharder-1"]
+      - wait_round:
+          shift: 100
+      - start: ["sharder-2"]
       - wait_round:
           shift: 200
-      - start: ["sharder-3"] # Will get the new LFB after 200 rounds
-      - wait_round:
-          shift: 100
       - wait_sharders_finalize_near_blocks:
-          sharders: ["sharder-1", "sharder-2", "sharder-3"]
+          sharders: ["sharder-1", "sharder-2"]
           timeout: "10m"
-
